+++ R/USGSm2m-package.R
@@ -4,6 +4,6 @@
 ## usethis namespace: start
 #' @importFrom R6 R6Class
 #' @importFrom rlang .data %||%
-#' @importFrom stats setNames
+#' @importFrom stats setNames na.omit
 ## usethis namespace: end
 NULL